backup/cloudrun_20250823_194502/backend/requirements.txt
flask==3.1.1
flask-cors==6.0.1
dnspython==2.7.0
aiodns==3.1.1
google-cloud-firestore==2.18.0
google-cloud-secret-manager==2.20.0
redis==5.0.1
python-dateutil==2.8.2
pandas==2.1.4
flask-limiter>=3.0.0
